refactor and cleanup, now mongodb work.
