+++ components/PreviewPicture.js
@@ -1,4 +1,4 @@
-import { StyledImage } from "./StyledImage";
+import { StyledImage } from "./styledComponents/StyledImage";
 
 export default function PreviewPicture({ imageSrc, imageName }) {
   return (
