+++ components/Text.js
@@ -1,4 +1,4 @@
-import { StyledText } from "./StyledText";
+import { StyledText } from "./styledComponents/StyledText";
 
 export default function Text({ texts }) {
   return <StyledText formula>{texts}</StyledText>;
