+++ components/styledComponents/StyledImage.js
