+++ components/styledComponents/StyledText.js
