+++ pages/ChooseFour/[picID].js
@@ -1,7 +1,7 @@
 import Text from "@/components/Text";
 import { texts } from "../../data/text";
 import FourPicsTable from "@/components/FourPicsTable";
-import { Container } from "@/components/Container";
+import { Container } from "@/components/styledComponents/Container";
 
 export default function ChooseFour() {
   return (
